Add "FlushAllOnReload" config option
Default value is "yes".

Even on --reload some runtime configuration was being retained. This was
a surprise to many users. Lets default to flushing all the runtime
configuration and also introduce a config option to go back the old
behavior of retaining; interface to zone assignments, and direct rules.

This also adjusts a few test cases that depend on the old
FlushAllOnReload=no behavior.

Fixes: #409